refactor: break components into separate folders
