+++ src/app/components/NavBar/TransparentNavBarBg.tsx
